добавляет отдельный модуль валидации
