+++ js/upload-form.js
@@ -1,10 +1,11 @@
+import { pristine, checkValidationForSubmitButton } from './validate-form.js';
+
 const uploadForm = document.querySelector('.img-upload__form');
 const uploadInput = uploadForm.querySelector('.img-upload__input');
 const uploadFormHashtags = uploadForm.querySelector('.text__hashtags');
 const uploadFormDescription = uploadForm.querySelector('.text__description');
 const uploadFormOverlay = uploadForm.querySelector('.img-upload__overlay');
 const uploadFormCloseButton = uploadForm.querySelector('.img-upload__cancel');
-const uploadFormSubmitButton = uploadForm.querySelector('.img-upload__submit');
 const uploadFormScaleSmaller = uploadForm.querySelector('.scale__control--smaller');
 const uploadFormScaleBigger = uploadForm.querySelector('.scale__control--bigger');
 const uploadFormScaleWrapper = uploadForm.querySelector('.img-upload__effect-level');
@@ -14,42 +15,11 @@ const uploadFormEffectLevel = uploadForm.querySelector('.effect-level__slider');
 const uploadFormEffectLevelInput = uploadForm.querySelector('.effect-level__value');
 const uploadFormEffectsList = uploadForm.querySelector('.effects__list');
 
-const MAX_QUANTITY_OF_HASHTAGS = 5;
-const DESCRIPTION_MAX_LENGTH = 140;
-
-let hashtagsErrors = [];
-let descriptionErrors = [];
-
-const hashtagsErrorsList = {
-  invalidHashtag: 'введён невалидный хэштег',
-  tooManyHashtags: `превышено количество хэштегов (${MAX_QUANTITY_OF_HASHTAGS})`,
-  noDoubling: 'хэштеги повторяются'
-};
-
-const descriptionErrorsList = {
-  tooLongText: `длина комментария больше ${DESCRIPTION_MAX_LENGTH} символов`
-};
-
-const pristine = new Pristine(uploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextTag: 'div',
-  errorTextClass: 'img-upload__field-wrapper--error'
-});
-
 const checkButtonPressForUploadForm = (evt) => {
   if (evt.key === 'Escape' && evt.target !== uploadFormHashtags && evt.target !== uploadFormDescription) {
     closeUploadForm();
   }
 };
-
-function toggleSubmitButton () {
-  if (pristine.validate()) {
-    uploadFormSubmitButton.disabled = false;
-  } else {
-    uploadFormSubmitButton.disabled = true;
-  }
-}
 
 function resetUploadFormFields () {
   uploadInput.value = '';
@@ -59,7 +29,7 @@ function resetUploadFormFields () {
 }
 
 function openUploadForm () {
-  toggleSubmitButton();
+  checkValidationForSubmitButton();
   uploadFormScaleWrapper.classList.add('hidden');
 
   uploadFormOverlay.classList.remove('hidden');
@@ -75,70 +45,11 @@ function closeUploadForm () {
   document.removeEventListener('keydown', checkButtonPressForUploadForm);
 }
 
-const validateHashtags = (string) => {
-  if (!string.trim()) {
-    return true;
-  }
-
-  let result = true;
-  hashtagsErrors = [];
-
-  const hashtagsArray = string.trim().toLowerCase().split(/\s+/).sort();
-  const hashtagRegExp = /^#[a-zа-яё0-9]{1,19}$/i;
-
-  hashtagsArray.forEach((hashtag) => {
-    if(!hashtagRegExp.test(hashtag)) {
-      result = false;
-      hashtagsErrors.push('invalidHashtag');
-    }
-  });
-
-  if (hashtagsArray.length > MAX_QUANTITY_OF_HASHTAGS) {
-    result = false;
-    hashtagsErrors.push('tooManyHashtags');
-  }
-
-  for (let i = 1; i < hashtagsArray.length; i++) {
-    if (hashtagsArray[i] === hashtagsArray[i - 1]) {
-      result = false;
-      hashtagsErrors.push('noDoubling');
-    }
-  }
-  return result;
-};
-
-const getErrorsMessage = (allErrorsList, foundErrorsList) => {
-  let errorMessage = '';
-  const uniqueErrors = [...new Set(foundErrorsList)];
-  uniqueErrors.forEach((error, index) => {
-    if (index > 0) {
-      errorMessage += ',<br>';
-    }
-    errorMessage += allErrorsList[error];
-  });
-  return errorMessage;
-};
-
-pristine.addValidator(uploadFormHashtags, validateHashtags, () => getErrorsMessage(hashtagsErrorsList, hashtagsErrors));
-
-const validateDescription = (string) => {
-  let result = true;
-  descriptionErrors = [];
-
-  if (string.length >= DESCRIPTION_MAX_LENGTH) {
-    result = false;
-    descriptionErrors.push('tooLongText');
-  }
-
-  return result;
-};
-
-pristine.addValidator(uploadFormDescription, validateDescription, () => getErrorsMessage(descriptionErrorsList, descriptionErrors));
-
 uploadInput.addEventListener('change', openUploadForm);
+
 uploadFormCloseButton.addEventListener('click', closeUploadForm);
 
-uploadForm.addEventListener('input', toggleSubmitButton);
+uploadForm.addEventListener('input', checkValidationForSubmitButton);
 
 uploadForm.addEventListener('submit', (evt) => {
   evt.preventDefault();
@@ -146,7 +57,6 @@ uploadForm.addEventListener('submit', (evt) => {
     uploadForm.submit();
   }
 });
-
 
 const SCALE_STEP = 25;
 const SCALE_MIN = 25;
